Add base relation for moengine
